requirements.txt
asgiref==3.2.10
autopep8==1.5.4
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
config==0.5.0.post0
dj-database-url==0.5.0
Django==3.1
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-tinymce==3.0.2
djangorestframework==3.11.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
python-dateutil==1.5
python-decouple==3.3
pytz==2020.1
requests==2.24.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
whitenoise==5.2.0
zipp==3.1.0
